changed id type from integer to long

+++ src/main/java/com/moviesAPI/moviesAPI/repositories/CharacterRepository.java
@@ -6,6 +6,6 @@ import com.moviesAPI.moviesAPI.entities.Character;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called characterRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface CharacterRepository extends CrudRepository<Character, Integer> {
+public interface CharacterRepository extends CrudRepository<Character, Long> {
 
 }
