+++ src/main/java/com/moviesAPI/moviesAPI/repositories/GenreRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called characterRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface GenreRepository extends CrudRepository<Genre, Integer> {
+public interface GenreRepository extends CrudRepository<Genre, Long> {
 
 }
